requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
Pillow>=10.0.0
numpy>=1.24.0
httpx>=0.26.0
ultralytics>=8.4.0
torch>=2.0.0
